+++ packages/http-client/src/errors/index.ts
@@ -1,3 +1,4 @@
 export { RequestError } from './request-error.js'
 export { ResponseError } from './response-error.js'
 export { ValidationError, InvalidDidError, MissingServiceEndpointError } from './validation-error.js'
+export { RequestTokenError, RequestTokenSigningError, RequestTokenVerificationError, RequestTokenMissingClaimsError, RequestTokenAudienceMismatchError } from './request-token-error.js'