+++ packages/http-client/tests/buffer-polyfill.cjs
@@ -1,0 +1,4 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const Buffer = require('buffer/').Buffer
+
+globalThis.Buffer = Buffer