+++ packages/http-client/tests/client.spec.ts
@@ -1,9 +1,20 @@
 import { expect } from 'chai'
-import { DidDhtMethod, DidKeyMethod } from '@web5/dids'
-import { TbdexHttpClient } from '../src/client.js'
-import { RequestError,ResponseError, InvalidDidError, MissingServiceEndpointError } from '../src/errors/index.js'
+import { DidDhtMethod, DidKeyMethod, PortableDid } from '@web5/dids'
+import { TbdexHttpClient, requestTokenRequiredClaims } from '../src/client.js'
+import {
+  RequestError,ResponseError,
+  InvalidDidError,
+  MissingServiceEndpointError,
+  RequestTokenMissingClaimsError,
+  RequestTokenAudienceMismatchError,
+  RequestTokenVerificationError,
+  RequestTokenSigningError
+} from '../src/errors/index.js'
 import { DevTools, Message, Rfq } from '@tbdex/protocol'
 import * as sinon from 'sinon'
+import { JwtHeaderParams, JwtPayload, PrivateKeyJwk, Secp256k1 } from '@web5/crypto'
+import { Convert } from '@web5/common'
+import { Jwt } from '@web5/credentials'
 
 const dhtDid = await DidDhtMethod.create({
   publish  : true,
@@ -90,7 +101,7 @@ describe('client', () => {
     })
   })
 
-  describe('getOfferings', async () => {
+  describe('getOfferings', () => {
     it('throws RequestError if service endpoint url is garbage', async () => {
       getPfiServiceEndpointStub.resolves('garbage')
       fetchStub.rejects({message: 'Failed to fetch on URL'})
@@ -140,7 +151,7 @@ describe('client', () => {
     })
   })
 
-  describe('getExchange', async () => {
+  describe('getExchange', () => {
     it('throws RequestError if service endpoint url is garbage', async () => {
       getPfiServiceEndpointStub.resolves('garbage')
       fetchStub.rejects({message: 'Failed to fetch on URL'})
@@ -190,7 +201,7 @@ describe('client', () => {
     })
   })
 
-  describe('getExchanges', async () => {
+  describe('getExchanges', () => {
     it('throws RequestError if service endpoint url is garbage', async () => {
       getPfiServiceEndpointStub.resolves('garbage')
       fetchStub.rejects({message: 'Failed to fetch on URL'})
@@ -240,7 +251,7 @@ describe('client', () => {
     })
   })
 
-  describe('getPfiServiceEndpoint', async () => {
+  describe('getPfiServiceEndpoint', () => {
     before(() => {
       getPfiServiceEndpointStub.restore()
       fetchStub.restore()
@@ -273,5 +284,104 @@ describe('client', () => {
       expect(serviceEndpoint).to.equal('https://localhost:9000')
     })
   })
-})
 
+  describe('generateRequestToken', () => {
+    let requesterPortableDid: PortableDid
+    before(async () => {
+      requesterPortableDid = await DidKeyMethod.create({ keyAlgorithm: 'secp256k1' })
+    })
+    it('throws a RequestTokenSigningError if requesterDid is not a valid PortableDid', async () => {
+      try {
+        await TbdexHttpClient.generateRequestToken({ requesterDid: {did: '', document: { id: '' }, keySet: {}}, pfiDid: '' })
+        expect.fail()
+      } catch (e) {
+        expect(e).to.be.instanceOf(RequestTokenSigningError)
+      }
+    })
+    it('includes all expected claims', async () => {
+      const requestToken = await TbdexHttpClient.generateRequestToken({ requesterDid: requesterPortableDid, pfiDid: 'did:key:1234' })
+      const decodedToken = await Jwt.verify({ jwt: requestToken })
+      expect(decodedToken.payload).to.have.all.keys(requestTokenRequiredClaims)
+    })
+    // TODO: decide if we want to ensure that the expiration date is not longer than 1 minute after the issuance date
+    it('sets expiration seconds to 1 minute after the time at which it was issued', async () => {
+      const requestToken = await TbdexHttpClient.generateRequestToken({ requesterDid: requesterPortableDid, pfiDid: 'did:key:1234' })
+      const decodedToken = await Jwt.verify({ jwt: requestToken })
+      expect(decodedToken.payload.exp - decodedToken.payload.iat).to.equal(60)
+    })
+  })
+
+  describe('verifyRequestToken', () => {
+    let pfiPortableDid: PortableDid
+    let header: JwtHeaderParams
+    let payload: JwtPayload
+
+    async function createRequestTokenFromPayload(payload) {
+      const privateKeyJwk = pfiPortableDid.keySet.verificationMethodKeys![0].privateKeyJwk
+      const base64UrlEncodedHeader = Convert.object(header).toBase64Url()
+      const base64UrlEncodedPayload = Convert.object(payload).toBase64Url()
+
+      const toSign = `${base64UrlEncodedHeader}.${base64UrlEncodedPayload}`
+      const toSignBytes = Convert.string(toSign).toUint8Array()
+      const signatureBytes = await Secp256k1.sign({ key: privateKeyJwk as PrivateKeyJwk, data: toSignBytes })
+      const base64UrlEncodedSignature = Convert.uint8Array(signatureBytes).toBase64Url()
+
+      return `${toSign}.${base64UrlEncodedSignature}`
+    }
+
+    before(async () => {
+      pfiPortableDid = await DidKeyMethod.create({ keyAlgorithm: 'secp256k1' })
+      header = { typ: 'JWT', alg: 'ES256K', kid: pfiPortableDid.document.verificationMethod![0].id }
+    })
+
+    beforeEach(() => {
+      payload = {
+        iat : Math.floor(Date.now() / 1000),
+        aud : pfiPortableDid.did,
+        iss : 'did:key:1234',
+        exp : Math.floor(Date.now() / 1000 + 60),
+        jti : 'randomnonce'
+      }
+    })
+
+    it('throws RequestTokenVerificationError if request token is not a valid jwt', async () => {
+      try {
+        await TbdexHttpClient.verifyRequestToken({ requestToken: '', pfiDid: pfiPortableDid.did })
+        expect.fail()
+      } catch(e) {
+        expect(e).to.be.instanceof(RequestTokenVerificationError)
+      }
+    })
+    it('throws RequestTokenMissingClaimsError if request token is missing any of the expected claims', async () => {
+      for (const claim of requestTokenRequiredClaims) {
+        const initialClaim = payload[claim]
+        try {
+          delete payload[claim]
+          const requestToken = await createRequestTokenFromPayload(payload)
+          await TbdexHttpClient.verifyRequestToken({ requestToken, pfiDid: pfiPortableDid.did })
+          expect.fail()
+        } catch(e) {
+          expect(e).to.be.instanceof(RequestTokenMissingClaimsError)
+          expect(e.message).to.include(`Request token missing ${claim} claim.`)
+        }
+        payload[claim] = initialClaim
+      }
+    })
+    it('throws RequestTokenAudienceMismatchError if aud claim does not match pfi did', async () => {
+      try {
+        payload.aud = 'squirtle'
+        const requestToken = await createRequestTokenFromPayload(payload)
+        await TbdexHttpClient.verifyRequestToken({ requestToken, pfiDid: pfiPortableDid.did })
+        expect.fail()
+      } catch(e) {
+        expect(e).to.be.instanceof(RequestTokenAudienceMismatchError)
+        expect(e.message).to.include('Request token contains invalid audience')
+      }
+    })
+    it('returns requester\'s DID if request token is valid', async () => {
+      const requestToken = await createRequestTokenFromPayload(payload)
+      const iss = await TbdexHttpClient.verifyRequestToken({ requestToken, pfiDid: pfiPortableDid.did })
+      expect(iss).to.equal('did:key:1234')
+    })
+  })
+})