+++ packages/protocol/src/message-kinds/rfq.ts
@@ -12,7 +12,7 @@ import Ajv from 'ajv'
  */
 export type CreateRfqOptions = {
   data: MessageKindModel<'rfq'>
-  metadata: Omit<MessageMetadata<'rfq'>, 'id' |'kind' | 'createdAt' | 'exchangeId'>
+  metadata: Omit<MessageMetadata<'rfq'>, 'id' | 'kind' | 'createdAt' | 'exchangeId'>
   private?: Record<string, any>
 }
 
@@ -25,7 +25,7 @@ export class Rfq extends Message<'rfq'> {
   readonly validNext = new Set<MessageKind>(['quote', 'close'])
 
   /** private data (PII or PCI) */
-  _private: Record<string, any>
+  _private: Record<string, any> | undefined
 
   /**
    * Creates an rfq with the given options
@@ -61,7 +61,7 @@ export class Rfq extends Message<'rfq'> {
    * @throws if {@link Rfq.payoutMethod} property `paymentDetails` cannot be validated against the provided offering's payoutMethod requiredPaymentDetails
    */
   async verifyOfferingRequirements(offering: Offering | ResourceModel<'offering'>) {
-    if (offering.metadata.id !== this.offeringId)  {
+    if (offering.metadata.id !== this.offeringId) {
       throw new Error(`offering id mismatch. (rfq) ${this.offeringId} !== ${offering.metadata.id} (offering)`)
     }
 
@@ -141,15 +141,19 @@ export class Rfq extends Message<'rfq'> {
    * @param offering - the offering to check against
    * @throws if rfq's claims do not fulfill the offering's requirements
    */
-  async verifyClaims(offering: Offering | ResourceModel<'offering'>) {
-    const credentials = PresentationExchange.selectCredentials(this.claims, offering.data.requiredClaims)
+  async verifyClaims(offering: Offering | ResourceModel<'offering'>): Promise<void> {
+    if (!offering.data.requiredClaims) {
+      return
+    }
+
+    const credentials = PresentationExchange.selectCredentials({ vcJwts: this.claims, presentationDefinition: offering.data.requiredClaims })
 
     if (!credentials.length) {
       throw new Error(`claims do not fulfill the offering's requirements`)
     }
 
     for (let credential of credentials) {
-      await VerifiableCredential.verify(credential)
+      await VerifiableCredential.verify({ vcJwt: credential })
     }
   }