+++ packages/protocol/src/resource-kinds/offering.ts
@@ -1,6 +1,5 @@
 import type { ResourceKindModel, ResourceMetadata } from '../types.js'
 import { Resource } from '../resource.js'
-import type { PresentationDefinitionV2 } from '@web5/credentials'
 
 /**
  * Options passed to {@link Offering.create}
@@ -66,9 +65,7 @@ export class Offering extends Resource<'offering'> {
   }
 
   /** Articulates the claim(s) required when submitting an RFQ for this offering. */
-  // TODO: Remove type annotation once type alias replaced with direct export in @web5/credentials
-  // [Link to the PR](https://github.com/TBD54566975/web5-js/pull/336)
-  get requiredClaims(): PresentationDefinitionV2 {
+  get requiredClaims() {
     return this.data.requiredClaims
   }
 }